+++ test_runner/src/main/kotlin/ftl/args/ArgsHelper.kt
@@ -117,11 +117,10 @@ object ArgsHelper {
 
         val bucketListOption = Storage.BucketListOption.prefix(bucket)
         val storageList = storage.list(bucketListOption).values?.map { it.name } ?: emptyList()
-        val targetBucket = storageList.find { it == bucket }
-        if (targetBucket != null) return targetBucket
+        if (storageList.contains(bucket)) return bucket
 
         return storage.create(
-            BucketInfo.newBuilder(targetBucket)
+            BucketInfo.newBuilder(bucket)
                 .setStorageClass(StorageClass.REGIONAL)
                 .setLocation(storageLocation)
                 .setLabels(bucketLabel)