+++ test_runner/src/main/kotlin/ftl/args/yml/FlankYml.kt
@@ -13,10 +13,13 @@ class FlankYmlParams(
     val smartFlankGcsPath: String = "",
 
     @field:JsonProperty("test-targets-always-run")
-    val testTargetsAlwaysRun: List<String> = emptyList()
+    val testTargetsAlwaysRun: List<String> = emptyList(),
+
+    @field:JsonProperty("files-to-download")
+    val filesToDownload: List<String> = emptyList()
 ) {
     companion object : IYmlKeys {
-        override val keys = listOf("testShards", "repeatTests", "smartFlankGcsPath", "test-targets-always-run")
+        override val keys = listOf("testShards", "repeatTests", "smartFlankGcsPath", "test-targets-always-run", "files-to-download")
     }
 
     init {