+++ test_runner/src/main/kotlin/ftl/util/ArtifactRegex.kt
@@ -1,7 +1,0 @@
-package ftl.util
-
-object ArtifactRegex {
-
-    val testResultRgx = Regex(".*test_result_\\d+\\.xml$")
-    val screenshotRgx = Regex(".*\\.png$")
-}